+++ packages/xchain-crypto/__tests__/crypto.test.ts
@@ -1,12 +1,4 @@
-import {
-  generatePhrase,
-  validatePhrase,
-  encryptToKeyStore,
-  decryptFromKeystore,
-  getPublicKeyPair,
-  getAddress,
-} from '../src/crypto'
-import { encodeAddress } from '../src/utils'
+import { generatePhrase, validatePhrase, encryptToKeyStore, decryptFromKeystore } from '../src/crypto'
 
 describe('Generate Phrase', () => {
   it('Generates 12-word phrase', () => {
@@ -39,29 +31,11 @@ describe('Validate Phrase', () => {
   })
 })
 
-describe('getPublicKeyPair', () => {
-  it('Generates Correct Publickeys from phrase', async () => {
-    const phrase = 'flush viable fury sword mention dignity ethics secret nasty gallery teach fever'
-    const publickeys = getPublicKeyPair(phrase)
-    expect(encodeAddress(publickeys.secp256k1?.getAddress() ?? '')).toEqual(getAddress(phrase))
-    expect(encodeAddress(publickeys.secp256k1?.getAddress() ?? '')).toEqual(
-      'thor13ymnnjc4jtymjsc297nulyf0d3qtdl8pyg0tvx',
-    )
-    expect(encodeAddress(publickeys.ed25519.getAddress())).toEqual('thor179wypy2zld6su560x25mj268dlscyqxhqfjp85')
-  })
-})
-
 describe('Export Keystore', () => {
   it('Generates Correct Keystore', async () => {
     const phrase = 'flush viable fury sword mention dignity ethics secret nasty gallery teach fever'
     const password = 'thorchain'
     const keystore = await encryptToKeyStore(phrase, password)
-    expect(encodeAddress(keystore.publickeys.secp256k1?.getAddress() ?? '')).toEqual(
-      'thor13ymnnjc4jtymjsc297nulyf0d3qtdl8pyg0tvx',
-    )
-    expect(encodeAddress(keystore.publickeys.ed25519.getAddress())).toEqual(
-      'thor179wypy2zld6su560x25mj268dlscyqxhqfjp85',
-    )
     expect(keystore.crypto.cipher).toEqual('aes-128-ctr')
     expect(keystore.crypto.kdf).toEqual('pbkdf2')
     expect(keystore.crypto.kdfparams.prf).toEqual('hmac-sha256')