+++ packages/xchain-crypto/CHANGELOG.md
@@ -3,6 +3,14 @@
 ### Update
 
 - Update dependencies
+- add `bip39.validatePhrase(phrase)` in `getSeed()`
+- add `bip39.validatePhrase(phrase)` in `encryptToKeystore()`
+
+### Breaking change
+
+- remove `getAddress()`
+- remove `getPublicKeyPair()`
+- remove pub keys from key store meta-data
 
 # v.0.2.2 (2020-11-23)